+++ tests/assertions/assert_cookie.test.js
@@ -22,48 +22,48 @@ describe("Assert Cookie", function() {
     });
 
     it("Assert Cookie", async() => {
-        await browser.assert.cookie("username");
-        await browser.assert.cookie("username", "arthur_dent");
+        await browser.assert.cookies("username");
+        await browser.assert.cookies("username", "arthur_dent");
     });
 
     it("Assert Cookie Throws", async() => {
         await utils.assertThrowsAssertionAsync(async() => {
-            await browser.assert.cookie("not-exists");
+            await browser.assert.cookies("not-exists");
         }, `Expected cookie "not-exists" to exist.`);
     });
 
     it("Assert Cookie Throws Invalid Value", async() => {
         await utils.assertThrowsAssertionAsync(async() => {
-            await browser.assert.cookie("username", "marvin");
+            await browser.assert.cookies("username", "marvin");
         }, `Expected cookie "username" to have value "marvin", "arthur_dent" found.`, "arthur_dent", "marvin");
     });
 
     it("Assert Cookie Throws Custom Message", async() => {
         await utils.assertThrowsAssertionAsync(async() => {
-            await browser.assert.cookie("username", "marvin", "cookie fails");
+            await browser.assert.cookies("username", "marvin", "cookie fails");
         }, `cookie fails`, "arthur_dent", "marvin");
     });
 
     it("Assert Not Cookie", async() => {
-        await browser.assert.not.cookie("not-exists");
-        await browser.assert.not.cookie("username", "marvin");
+        await browser.assert.not.cookies("not-exists");
+        await browser.assert.not.cookies("username", "marvin");
     });
 
     it("Assert Not Cookie Throws", async() => {
         await utils.assertThrowsAssertionAsync(async() => {
-            await browser.assert.not.cookie("username");
+            await browser.assert.not.cookies("username");
         }, `Expected cookie "username" to not exist.`);
     });
 
     it("Assert Not Cookie Throws Invalid Value", async() => {
         await utils.assertThrowsAssertionAsync(async() => {
-            await browser.assert.not.cookie("username", "arthur_dent");
+            await browser.assert.not.cookies("username", "arthur_dent");
         }, `Expected cookie "username" to not have value "arthur_dent".`);
     });
 
     it("Assert Not Cookie Throws Custom Message", async() => {
         await utils.assertThrowsAssertionAsync(async() => {
-            await browser.assert.not.cookie("username", "arthur_dent", "not cookie fails");
+            await browser.assert.not.cookies("username", "arthur_dent", "not cookie fails");
         }, `not cookie fails`);
     });
 });