+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * Mock assert postBody
   * Logs parse js objects correctly in console module
   * Puppeteer updated to 1.9.0
+  * Browser.assert.cookie and browser.assert.webworker deprecated in favor of cookies and webworkers
   * Other dependencies updated
   * Minor improvements in mock handling
   * Minor code improvements on browserFactory and Modules
@@ -174,7 +175,7 @@
 0.4.6 / 2018-03-24
 ==================
 
-  * Browser.assert.cookie
+  * Browser.assert.cookies
   * Cookies submodule
   * Remove default assertions errors